+++ NEWS.md
@@ -1,3 +1,7 @@
+# yulab.utils 0.0.5.001
+
++ `str_starts()` and `str_ends()` (2022-07-29, Fri)
+
 # yulab.utils 0.0.5
 
 + `get_fun_from_pkg()` outputs friendly message if the pkg is not installed (2022-06-08, Wed)